+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Optimize chain calls in `AuthorizationService`. (#228)
 - Fix licence headers in some source files. (#232)
 - Move `EthereumCredentials` class to `com.iexec.sms.tee.challenge` package. (#234)
+- Use `@DataJpaTest` to run secrets and TEE challenges tests on H2 database. (#236)
 
 ## [[8.4.0]](https://github.com/iExecBlockchainComputing/iexec-sms/releases/tag/v8.4.0) 2024-01-10